refactor: split the shared dependency providers along domain lines
geometrikks/api is gone: the last shared module, api/dependencies.py,
splits into domain-owned providers (security keeps the CrowdSec
service/poller, enrichment repo, and the in-memory pagination provider;
analytics gets its stats repos; system gets the ingestion provider) and
create_settings_provider moves to server/dependencies.py next to the
composition root that registers it.
