+++ geometrikks/domain/analytics/controllers.py
@@ -38,7 +38,7 @@ from geometrikks.domain.analytics.dtos import (
 )
 from geometrikks.domain.analytics.repositories import StatsGranularity, get_stats_granularity
 
-from geometrikks.api.dependencies import (
+from geometrikks.domain.analytics.dependencies import (
     provide_live_stats_repo,
     provide_summary_stats_repo
 )
