+++ geometrikks/domain/security/dependencies.py
@@ -1,4 +1,4 @@
-"""Shared dependency providers for API layer."""
+"""Dependency providers for the security (CrowdSec) domain."""
 from __future__ import annotations
 
 from typing import Annotated
@@ -6,39 +6,14 @@ from typing import Annotated
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from litestar import Request
-from litestar.di import NamedDependency, Provide
+from litestar.di import NamedDependency
 from litestar.params import QueryParameter
 from advanced_alchemy.extensions.litestar import filters
 
-from geometrikks.config.settings import Settings
 from geometrikks.server import runtime
 from geometrikks.services.crowdsec import CrowdSecService
 from geometrikks.services.crowdsec.stream import CrowdSecStreamPoller
-from geometrikks.services.ingestion import LogIngestionService
-from geometrikks.domain.analytics.repositories import LiveStatsRepository, SummaryStatsRepository
 from geometrikks.domain.security.repositories import SecurityEnrichmentRepository
-
-
-def create_settings_provider(settings: Settings) -> Provide:
-    """Build the app-level ``settings`` dependency around an explicit object.
-
-    ``create_app()`` registers this so request handlers receive the exact
-    settings the app was composed with; tests can pass their own ``Settings``
-    to ``create_app(settings=...)`` instead of mutating process state.
-    """
-
-    def provide_settings() -> Settings:
-        return settings
-
-    return Provide(provide_settings, sync_to_thread=False)
-
-
-def provide_ingestion_service(request: Request) -> LogIngestionService | None:
-    """Provide the LogIngestionService from app state.
-
-    Returns None if the service is not available (degraded mode).
-    """
-    return runtime.get_ingestion_service(request.app)
 
 
 def provide_crowdsec_service(request: Request) -> CrowdSecService | None:
@@ -82,16 +57,3 @@ def provide_limit_offset_pagination(
         Number of items per page.
     """
     return filters.LimitOffset(page_size, page_size * (current_page - 1))
-
-
-async def provide_summary_stats_repo(
-    db_session: NamedDependency[AsyncSession],
-) -> SummaryStatsRepository:
-    """Provide SummaryStatsRepository for querying summary statistics."""
-    return SummaryStatsRepository(session=db_session)
-
-async def provide_live_stats_repo(
-    db_session: NamedDependency[AsyncSession],
-) -> LiveStatsRepository:
-    """Provide LiveStatsRepository for querying raw data tables."""
-    return LiveStatsRepository(session=db_session)
